refactor(memory): relocate bounded daemon calls
